Patch from Dmitry: addresses suspicious casts - casts where a pointer size type is casted to a 32-bit integer and then is implicitly converted back to a pointer size type.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@21928 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/Commands/CommitCommand.cpp
@@ -52,7 +52,7 @@ void CommitCommand::InitProgressDialog
     progDlg.SetCommitMessage(commitDlg.m_sLogMessage);
     progDlg.SetDepth(commitDlg.m_bRecursive ? svn_depth_infinity : svn_depth_empty);
     progDlg.SetSelectedList(commitDlg.m_selectedPathList);
-    progDlg.SetItemCount(static_cast<long>(commitDlg.m_itemsCount));
+    progDlg.SetItemCount(commitDlg.m_itemsCount);
     progDlg.SetBugTraqProvider(commitDlg.m_BugTraqProvider);
     progDlg.SetRevisionProperties(commitDlg.m_revProps);
 }
